Version 1 complete

+++ app/controllers/CommentsController.php
@@ -21,13 +21,13 @@ class CommentsController extends \BaseController {
 	{
 		 $validate = Comment::validate(Input::all());
 
-        if ( $validate->passes() ) {
+        if ( $validate->passes() && Auth::check()) {
 			$comment = new Comment;
 			$comment->content = String::makeLinks( HTML::entities(Input::get('content')) );
-			$comment->author_id ='1';
+			$comment->author_id = Auth::user()->id;
 			$comment->gist_id = Gist::whereGistUri($uri)->firstOrFail()->id;
 			if($comment->save()){
-				return Redirect::back()
+				return Redirect::to('gist/'.$uri.'#comments')
 				->with('message', 'Comment posted successfully!');
 			}else{
 				return Redirect::back()
@@ -41,49 +41,6 @@ class CommentsController extends \BaseController {
 	}
 
 	/**
-	 * Store a newly created resource in storage.
-	 *
-	 * @return Response
-	 */
-	public function store()
-	{
-		//
-	}
-
-	/**
-	 * Display the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function show($id)
-	{
-		//
-	}
-
-	/**
-	 * Show the form for editing the specified resource.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function edit($id)
-	{
-		//
-	}
-
-	/**
-	 * Update the specified resource in storage.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function update($id)
-	{
-		//
-	}
-
-	/**
 	 * Remove the specified resource from storage.
 	 *
 	 * @param  int  $id
@@ -92,15 +49,20 @@ class CommentsController extends \BaseController {
 	public function destroy($id)
 	{
 		$comment = Comment::findOrFail($id);
-		if($comment->delete())
+		if (Auth::user()->id == $comment->author_id)
 		{
-	  		return Redirect::back()
-      	->with('message', 'Successfully deleted comment!');
-		}
-		else{
+			if($comment->delete())
+			{
+		  		return Redirect::back()
+	      	->with('message', 'Successfully deleted comment!');
+			}
+
 			return Redirect::back()
 			->with('error', 'Could not delete comment');
 		}
+		
+		return Redirect::back()
+		->with('error', 'Could not delete comment');
 	}
 
 }
